refactor: move autocomplete index to canonical package
